Fix endianness issues
When running on little-endian hosts, STD/STS/STX/STY would invert endianness
and various pushes/pulls to/from the stack in PSH/PUL and subroutine call and
return opcodes would also invert endianness (though this was less important
since such data would usually be accessed with a matchingly buggy opcode that
inverted it back again).

The definitions of ACCUM_A and ACCUM_B used to assume the layout of union
reg_accum was consistent with a little-endian host; now we adapt as necessary.