Fix signed branch conditions
GE, LT, GT and LE were not correctly honoured in the case when both N and V
were set (e.g. due to comparison of a positive accumulator against a
negative operand when the difference between them exceeded 0x8000 for 16-bit
or 0x80 for 8-bit operations). This was due to the use of the `^` bitwise
XOR operator for what should have been a logical XOR operation. It's only
necessary to use a single rather than double `!` to normalise the inputs to
the XOR since its output is invariant if you invert both its inputs.